requirements.txt
streamlit>=1.29.0
pandas>=1.5.3
numpy>=1.24.3
matplotlib>=3.7.1
seaborn>=0.12.2
streamlit-pandas-profiling>=0.1.3
sweetviz>=2.1.4
pandas-profiling>=3.2.0
openpyxl>=3.1.2
nltk>=3.8.1
langdetect>=1.0.9
opencv-python>=4.9.0.80
Pillow>=10.3.0
scikit-learn>=1.4.2
imagehash>=4.3.1
tqdm>=4.66.2
pydub>=0.25.1
librosa>=0.10.1
soundfile>=0.12.1
scipy>=1.10.1
speechbrain>=0.5.16
sentencepiece>=0.1.99
torch>=2.0.1